created useAvailableTools
